Add example for federated gateway
Adds client credentials source, with caching for if a token
is requested several times before expiry.

GetFunction was returning FunctionDeployment, which is the input
to deploy a function, not the output from a function's status.

That meant there was no way to obtain the name, replicas etc.

Tested with federated gateway authentication source.

Add clock skew on expiry and lock for read/write

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>