Use function name as cache key for function access tokens
Remove the source token from the cache key.

By using the source token as part of the cache key the cache would be
invalidated when the original access token had expired which is confusing
for users and leads to higher load on the token exchange endpoint if
short lived tokens are used.

Signed-off-by: Han Verstraete (OpenFaaS Ltd) <han@openfaas.com>